requirements.txt
Django==4.2
# psycopg2-binary==2.9.6
# celery==5.2.7
# django-celery-results==2.5.0
# redis==4.5.4
# django-redis==5.2.0
# django-celery-beat==2.5.0
djangorestframework==3.14.0
Pillow==9.5.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
